Tell It's All Text! that file textarea is markdown.
The [It's All Text!](https://github.com/docwhat/itsalltext) Firefox
extension lets you edit textareas in your favorite desktop editor.

This change tells IAT that it should use an extension of `.md`
when creating the file for the desktop editor; which allows
syntax highlighting and indenting to work in most editors.